Remove extra -%} in kafka.properties.
The extra -%} is copied as a literal in the final kafka.properties file. That results in a config named -%} being created. This config is problematic for REST Proxy, because REST Proxy assumes all config names are URL safe, therefore it tries to construct an URL with this config name on it and it fails.